feat(auth): get code and redirect back to activity
Here we have the implementation to login and receive the callback code
parameter in AuthReceiverActivity.

Compose is suggesting single activity application but a standalone
activity may suit for now, as the main activity is already running where
we can not get intent parameters in `onCreate`, it's fine.

This commit also makes the app build differ in debug mode and release
mode.